darpan-ai-engine/requirements.txt
Flask
Flask-Cors
gunicorn
vertexai
google-cloud-aiplatform
google-cloud-discoveryengine